Move listener to MetricClient.scala
